move python wrapper for schedules to their own files
