+++ .ycm_extra_conf.py
@@ -51,6 +51,10 @@ flags = [
     '-I',
     'arbor',
     '-I',
+    'arbor/include',
+    '-I',
+    'build/arbor/include',
+    '-I',
     'ext/json/single_include',
     '-I',
     'build/include',
