+++ python/event_generator.hpp
@@ -3,8 +3,6 @@
 #include <arbor/common_types.hpp>
 #include <arbor/schedule.hpp>
 
-#include <pybind11/pybind11.h>
-
 namespace pyarb {
 
 struct event_generator_shim {
@@ -12,8 +10,8 @@ struct event_generator_shim {
     double weight;
     arb::schedule time_sched;
 
-    event_generator_shim(arb::cell_member_type cell, double event_weight, arb::schedule sched):
-        target(cell),
+    event_generator_shim(arb::cell_member_type gid, double event_weight, arb::schedule sched):
+        target(gid),
         weight(event_weight),
         time_sched(std::move(sched))
     {}
