+++ python/pyarb.cpp
@@ -10,10 +10,12 @@ void register_config(pybind11::module& m);
 void register_contexts(pybind11::module& m);
 void register_event_generators(pybind11::module& m);
 void register_identifiers(pybind11::module& m);
+void register_recipe(pybind11::module& m);
+void register_schedules(pybind11::module& m);
+
 #ifdef ARB_MPI_ENABLED
 void register_mpi(pybind11::module& m);
 #endif
-void register_recipe(pybind11::module& m);
 }
 
 PYBIND11_MODULE(arbor, m) {
@@ -28,4 +30,5 @@ PYBIND11_MODULE(arbor, m) {
     pyarb::register_mpi(m);
     #endif
     pyarb::register_recipe(m);
+    pyarb::register_schedules(m);
 }
